devops: restructure LO structure
This commit restructures the DevOps _Learning Outcome_ structure. It sets up the initial skeleton structure for the LOs for _Junior DevOps Engineer_; with the first 10 LOs.

The approach is to make DevOps training more platform agnostic, since the previous one was based on AWS. The learning outcomes whoever will be implemented through AWS platform.

Ref: issue #7