Note
Active Record is one of the seven Ruby gems working harmoniously together to make up the rails framework. It takes care of all the database related stuff. It's known as an "ORM". ORM stands for Object-Relational-Mapping. It basically means that Active Record takes data which is stored in a database table using rows and columns, which needs to be modified or retrieved by writing SQL statements (if you're using a SQL database), and it lets you interact with that data as though it was a normal Ruby object.