Skill Description edits
Removed:  Intents bind activities together at runtime. An application usually consists of multiple activities that are loosely bound to each other. Typically, one activity in an application is specified as the "main" activity, which is presented to the user when launching the application for the first time.  Each time a new activity starts, the previous activity is stopped, but the system preserves the activity in a stack (the "back stack").